Hotfix(SignUpScreen): fix go back button & UI refactors
