--- src/Loading.js
@@ -1,27 +1,0 @@
-import React, { Component } from "react";
-import { View, Text, ActivityIndicator, StyleSheet } from "react-native";
-import firebase from "react-native-firebase";
-
-export default class Loading extends Component {
-  componentDidMount() {
-    firebase.auth().onAuthStateChanged(user => {
-      this.props.navigation.navigate(user ? "MainTabNavigator" : "Login");
-    });
-  }
-  render() {
-    return (
-      <View style={styles.container}>
-        <Text>Loading</Text>
-        <ActivityIndicator size="large" />
-      </View>
-    );
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center"
-  }
-});
